fix: Display the incoming hp correctly
The bug was caused by the incorrect module of the cursor. In fact, the user cursor was module only for showing the grid, but the cursor coordinates were not clamped. Since the incoming hp were calculated accessing the cell pointed by the cursor, but the cursor coordinates were not clamped, the cell pointed differed from the one showed and the hp calculation resulted erroneous.